docker: add Dockerfile and bash wrapper for Umbrel app
This commit adds a Dockerfile and wrapper script that allows a chantool
image to be used as an Umbrel app for easy use by non-technical users.
The "app" only spins up a background container that can be executed into
and run CLI commands manually.
The entrypoint and bash wrapper scripts are only there to give guidance
and improve UX.